+++ src/pages/ProductDetailPage/index.tsx
@@ -6,7 +6,7 @@ import BottomButton from '@/components/BottomButton';
 import SectionDivider from '@/components/SectionDivider';
 import styled from '@emotion/styled';
 import { fetchProductSummary } from '@/api/products';
-import { Heart } from 'lucide-react';
+import WishButton from './WishButton';
 
 interface ProductSummaryData {
   id: number;
@@ -62,10 +62,7 @@ const ProductDetailPage = () => {
         </SectionContent>
       </Container>
 
-      <WishIcon>
-        <StyledHeartIcon />
-        <Count>123</Count>
-      </WishIcon>
+      <WishButton />
       <BottomButton>주문하기</BottomButton>
     </Layout>
   );
@@ -131,35 +128,4 @@ const SectionContent = styled.div`
   min-height: 200px;
   border-radius: 8px;
   padding: 16px;
-`;
-
-const WishIcon = styled.button`
-  position: fixed;
-  bottom: 72px;
-  left: 16px;
-  background: white;
-  border: none;
-  border-radius: 50%;
-  width: 48px;
-  height: 48px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-  font-size: 16px;
-  cursor: pointer;
-  padding: 4px 0;
-`;
-
-const StyledHeartIcon = styled(Heart)`
-  color: ${({ theme }) => theme.color.gray600};
-  width: 18px;
-  height: 18px;
-`;
-
-const Count = styled.div`
-  font-size: 10px;
-  color: ${({ theme }) => theme.color.gray600};
-  line-height: 1;
 `;
